Move data links suggestions to grafana-data
