+++ packages/grafana-data/src/index.ts
@@ -7,3 +7,4 @@ export * from './datetime';
 export * from './text';
 export * from './valueFormats';
 export * from './field';
+export * from './links';
