+++ packages/grafana-ui/src/utils/dataLinks.ts
@@ -1,21 +1,6 @@
 import { ContextMenuItem } from '../components/ContextMenu/ContextMenu';
 import { LinkModelSupplier } from '@grafana/data';
 
-export const DataLinkBuiltInVars = {
-  keepTime: '__url_time_range',
-  timeRangeFrom: '__from',
-  timeRangeTo: '__to',
-  includeVars: '__all_variables',
-  seriesName: '__series.name',
-  fieldName: '__field.name',
-  valueTime: '__value.time',
-  valueNumeric: '__value.numeric',
-  valueText: '__value.text',
-  valueRaw: '__value.raw',
-  // name of the calculation represented by the value
-  valueCalc: '__value.calc',
-};
-
 /**
  * Delays creating links until we need to open the ContextMenu
  */
