+++ public/app/features/dashboard/panel_editor/GeneralTab.tsx
@@ -8,9 +8,8 @@ import './../../panel/GeneralTabCtrl';
 
 // Types
 import { PanelModel } from '../state/PanelModel';
-import { DataLink } from '@grafana/data';
+import { DataLink, getPanelLinksVariableSuggestions } from '@grafana/data';
 import { PanelOptionsGroup, DataLinksEditor } from '@grafana/ui';
-import { getPanelLinksVariableSuggestions } from 'app/features/panel/panellinks/link_srv';
 
 interface Props {
   panel: PanelModel;
