+++ public/app/plugins/datasource/loki/configuration/DerivedField.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { css } from 'emotion';
-import { Button, FormField, VariableSuggestion, DataLinkInput, stylesFactory } from '@grafana/ui';
+import { VariableSuggestion } from '@grafana/data';
+import { Button, FormField, DataLinkInput, stylesFactory } from '@grafana/ui';
 import { DerivedFieldConfig } from '../types';
 
 const getStyles = stylesFactory(() => ({
